pyproject.toml
[project]
name = "content-aggregator-bot"
version = "0.1.0"
description = "soon..."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiogram",
    "aiohttp",
    "aiosqlite",
    "loguru",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "pyyaml",
    "sqlalchemy",
    "telethon",
    "vk-api",
]

[dependency-groups]
dev = ["ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
bot = "app.__main__:run_main"
